requirements.txt
flask
ultralytics
Pillow
numpy
opencv-python